Split test cases
