+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/UriShouldNotBeHardcoded.cs
@@ -1,8 +1,5 @@
 using System;
 using System.IO;
-using System.Web;
-using System.Web.Mvc;
-using static System.Web.HttpContext;
 
 namespace Tests.Diagnostics
 {
@@ -63,62 +60,4 @@ namespace Tests.Diagnostics
             var concat1 = s + "\\" + s;
         }
     }
-
-    #region virtual-path-tests
-
-    public class AspDotNetController : Controller
-    {
-        public ActionResult Foo()
-        {
-            string virtualPath = "~/scripts/relative.js"; // Noncompliant FP
-            Server.MapPath(virtualPath);
-
-            // System.Web.Mvc.HttpServerUtilityBase
-            Server.MapPath("~/scripts/relative.js");
-            Server.MapPath("/scripts/absolute.js");
-            Server.MapPath(@"C:\path\stuff.txt"); // Noncompliant (n.b. method will throw exception)
-
-            // System.Web.Mvc.HttpRequestBase
-            Request.MapPath("~/scripts/relative.js");
-            Request.MapPath("/scripts/absolute.jss");
-            Request.MapPath(@"C:\path\stuff.txt"); // Noncompliant (n.b. method will throw exception)
-
-            // System.Web.Mvc.HttpResponseBase
-            Response.ApplyAppPathModifier("~/scripts/relative.js");
-            Response.ApplyAppPathModifier("/scripts/absolute.js");
-            Response.ApplyAppPathModifier(@"C:\path\stuff.txt"); // Noncompliant
-
-            // System.Web.VirtualPathUtility (all methods)
-            VirtualPathUtility.GetDirectory("~/scripts/relative.js");
-            VirtualPathUtility.IsAppRelative("/scripts/absolute.js");
-            VirtualPathUtility.GetExtension("/scripts/absolute.js");
-            // ...
-            VirtualPathUtility.GetFileName(@"C:\path\stuff.txt"); // Noncompliant (n.b. method will throw exception)
-            VirtualPathUtility.Combine("root", @"C:\path\stuff.txt"); // Noncompliant
-            // ...
-
-            // System.Web.Mvc.UrlHelper
-            UrlHelper urlHelper = new UrlHelper(Current.Request.RequestContext);
-            urlHelper.Content("~/scripts/relative.js");
-            urlHelper.Content("/scripts/absolute.js");
-            urlHelper.Content(@"C:\path\stuff.txt"); // Noncompliant
-
-            return View();
-        }
-    }
-
-    public class DotNetCoreController : Microsoft.AspNetCore.Mvc.Controller
-    {
-        public Microsoft.AspNetCore.Mvc.IActionResult Bar(Microsoft.AspNetCore.Routing.IRouter router)
-        {
-            new Microsoft.AspNetCore.Mvc.VirtualFileResult(@"C:\path\stuff.txt", "text/javascript"); // Noncompliant
-            new Microsoft.AspNetCore.Routing.VirtualPathData(router, @"C:\path\stuff.txt"); // Noncompliant
-
-            new Microsoft.AspNetCore.Mvc.VirtualFileResult("/scripts/file.js", "text/javascript");
-            new Microsoft.AspNetCore.Routing.VirtualPathData(router, "/my/path");
-            return View();
-        }
-    }
-
-    #endregion
 }
